@@ -19,35 +19,41 @@ classifiers =
     Topic :: System :: Hardware :: Hardware Drivers
     Topic :: Software Development :: Libraries :: pygame
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.11
 
 [options]
 zip_safe = False
-packages = find:
-python_requires = >=3.6, <4
-namespace_packages = luma
+packages = find_namespace:
+python_requires = >=3.7, <4
 include_package_data = True
 install_requires =
-    luma.core>=2.2.0
+    luma.core>=2.4.0
     pygame
 tests_require =
     pytest
     pytest-cov
     pytest-timeout
 
+[options.packages.find]
+include = luma*
+
 [options.extras_require]
 docs = sphinx>=1.5.1
-qa = flake8; rstcheck
-test = pytest; pytest-cov; pytest-timeout
+qa =
+    flake8
+    rstcheck
+test =
+    pytest
+    pytest-cov
+    pytest-timeout
 
 [options.package_data]
 luma.emulator.images =
-    luma/emulator/images/led_on.png
-    luma/emulator/images/led_off.png
-    luma/emulator/images/7-segment.png
+    *.png
 
 [bdist_wheel]
 universal = 1
